+++ src/main/java/io/anserini/search/InvertedDenseSearcher.java
@@ -1,0 +1,4 @@
+package io.anserini.search;
+
+public class InvertedDenseSearcher {
+}
