--- src/main/java/io/anserini/search/SimpleInvertedDenseSearcher.java
@@ -1,4 +1,0 @@
-package io.anserini.search;
-
-public class SimpleInvertedDenseSearcher {
-}
